Moved FPS definition in sprite sheet XML file.

+++ src/scene/components/SSceneComSprite.h
@@ -41,22 +41,22 @@ public:
 
     /* Get current sprite frame. 
      * Always returns 0 if texture is not animated */
-    uint32_t getCurrentFrame() const;
+    float getCurrentFrame() const;
     
     /* Increases animation to next frame. 
      * Multiple frames can be ticked with by setting argument @count.
      * Clamps to begining if needed */
-    void setCurrentFrame(uint32_t count);
+    void setCurrentFrame(float count);
     
     /* Returns how long one sprite frame lasts. (animation speed - fps/1000.0) */
-    float getFrameTime() const;
+    uint32_t getFrameTime() const;
     
     /* Returns time when was the last frame changed */
-    uint64_t getLastFrameTime() const;
+    uint32_t getLastFrameTime() const;
 
     /* Sets value to @_LastFrameTime. 
      * If @time == 0 (default argumet) current time will be set */
-    void setLastFrameTime(uint64_t time);
+    void setLastFrameTime(uint32_t time);
 
     /* Returns reference texture part (rect) that should be rendered.
      * If sprite is animated, this rect will be current animation frame,
@@ -78,13 +78,12 @@ private:
     /* Reference to sprite sheet for faster access */
     SComSpriteSheet* _SpriteSheet;
 
-    /* How many 'frames pre second' this sprite has.
-     * Default 0.0f */
-    float _FrameTime;
+    /* Duration of one frame in milisecounds */
+    uint32_t _FrameTime;
 
     /* Index of frame that is currently been drawn */
     uint32_t _CurrentFrame;
     
-    /* Time in nanoseconds when the last frame was changed */
-    uint64_t _LastFrameTime;
+    /* Time in milisecounds when the last frame was changed */
+    uint32_t _LastFrameTime;
 };
